#20183 move current problem contribution to editor status
